backend/app/database.py
from supabase import create_client
from dotenv import load_dotenv
import os

load_dotenv()

# Get Supabase credentials from environment variables
SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

# Create Supabase client
supabase = create_client(SUPABASE_URL, SUPABASE_KEY)

def get_db():
    """Return Supabase client"""
    return supabase