backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
emails==0.6
jinja2==3.1.3
python-dateutil==2.8.2
aiohttp==3.9.3
requests==2.31.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.123.0
openai==1.12.0
python-dotenv==1.0.0
anthropic-bedrock>=0.2.0